refactor(cockpit): extract section renderer
